Shuffling classes in sklearn.preprocessing
