+++ bdranalytics/sklearn/preprocessing/__init__.py
@@ -1,0 +1,7 @@
+from .encoding import WeightOfEvidenceEncoder, StringIndexer, LeaveOneOutEncoder
+from .scaling import ScaledRegressor
+
+__all__ = ['ScaledRegressor',
+           'WeightOfEvidenceEncoder',
+           'StringIndexer',
+           'LeaveOneOutEncoder']
