+++ bdranalytics/sklearn/preprocessing/encoding.py
@@ -1,6 +1,7 @@
 import numpy as np
 import pandas as pd
-from sklearn.base import BaseEstimator, TransformerMixin
+from sklearn.base import BaseEstimator
+from sklearn.base import TransformerMixin
 
 
 class WeightOfEvidenceEncoder(BaseEstimator, TransformerMixin):
@@ -61,7 +62,7 @@ class WeightOfEvidenceEncoder(BaseEstimator, TransformerMixin):
                 print("computing weight of evidence for column {:s}".format(col))
 
             stat[col] = ((df.groupby(col)[y_col_index].agg(compute_bucket_woe)
-                         + np.log(total_negative / total_positive)).to_dict())
+                          + np.log(total_negative / total_positive)).to_dict())
 
         self.stat = stat
 
@@ -83,8 +84,8 @@ class WeightOfEvidenceEncoder(BaseEstimator, TransformerMixin):
             stat = pd.DataFrame.from_dict(self.stat[col], orient='index')
 
             ser = (pd.merge(df, stat, left_on=col, right_index=True, how='left')
-                   .sort_index()
-                   .reindex(df.index))[0]
+                .sort_index()
+                .reindex(df.index))[0]
 
             # fill missing values with
             if self.verbose:
@@ -99,21 +100,6 @@ class WeightOfEvidenceEncoder(BaseEstimator, TransformerMixin):
             out = df
 
         return out
-
-
-class ColumnSelector(BaseEstimator, TransformerMixin):
-    def __init__(self, columns):
-        self.columns = columns
-
-    def fit(self, X, y=None):
-        return self
-
-    def transform(self, X):
-        try:
-            return X[self.columns]
-        except:
-            print("Could not find selected columns {:s} in available columns {:s}".format(self.columns, X.columns))
-            raise
 
 
 class StringIndexer(BaseEstimator, TransformerMixin):
